Adicionando campos 'profit_in_money' e 'stake' em 'bet'

+++ backend/controllers/bet_controller.py
@@ -1,5 +1,6 @@
 from fastapi import APIRouter, Depends
 from sqlalchemy.orm import Session
+from decimal import Decimal
 
 from database import get_db
 from schemas.bet_request import BetRequest
@@ -17,5 +18,5 @@ def get_by_id(id: int, db: Session = Depends(get_db)):
     return service.get_by_id(id, db)
 
 @router.put("/bet/{id}/exit/{exit_odd}")
-def exit_bet(id: int, exit_odd: float, db: Session = Depends(get_db)):
+def exit_bet(id: int, exit_odd: Decimal, db: Session = Depends(get_db)):
     return service.exit(id, exit_odd, db)
